requirements.txt
weaviate-client
numpy
scikit-learn
requests
pydantic
weaviate-client-python